Rename demo -> appdemo
