+++ app/src/androidTest/java/com/talkable/appdemo/ApplicationTest.java
@@ -1,4 +1,4 @@
-package com.talkable.demo;
+package com.talkable.appdemo;
 
 import android.app.Application;
 import androidx.test.core.app.ApplicationProvider;
