+++ app/src/main/java/com/talkable/appdemo/DemoApp.java
@@ -1,4 +1,4 @@
-package com.talkable.demo;
+package com.talkable.appdemo;
 
 import android.app.Application;
