+++ app/src/main/java/com/talkable/appdemo/MainActivity.java
@@ -1,4 +1,4 @@
-package com.talkable.demo;
+package com.talkable.appdemo;
 
 import android.app.Activity;
 import android.os.Bundle;
