+++ app/src/main/res/layout/activity_main.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.talkable.demo.MainActivity">
+    tools:context="com.talkable.appdemo.MainActivity">
 
     <ScrollView
         android:layout_width="fill_parent"
