+++ app/src/test/java/com/talkable/appdemo/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.talkable.demo;
+package com.talkable.appdemo;
 
 import org.junit.Test;
